+++ config/medium/Snowflakes/code/CGCode.js
@@ -1,0 +1,12 @@
+var inputs = readline().split(' ');
+const h = parseInt(inputs[0]);
+const w = parseInt(inputs[1]);
+for (let i = 0; i < h; i++) {
+    const row = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('Number of snowflakes');
+console.log('Number of unique snowflakes');