+++ config/medium/Snowflakes/code/CGCode.ts
@@ -1,0 +1,12 @@
+var inputs: string[] = readline().split(' ');
+const h: number = parseInt(inputs[0]);
+const w: number = parseInt(inputs[1]);
+for (let i = 0; i < h; i++) {
+    const row: string = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('Number of snowflakes');
+console.log('Number of unique snowflakes');