+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Lines Intersections".
 - configuration for "Tic Tac Toe Engine".
 - configuration for "Shortest Palindromic Path".
+- configuration for "Snowflakes".
 
 ## [1.21.0] - 2025-11-30
 ### Added